+++ apps/gimli/changelog.md
@@ -4,6 +4,11 @@ _(Note: This changelog focuses on the major changes between the different
 versions. Therefore, it may not contain all changes. Especially smaller fixes or
 improvements may be omitted.)_
 
+## Version 0.2.0-pre (2022-03-22)
+
+* Hashing now uses difference hashing only.
+* Results are sorted by similarity, starting with highest values.
+
 ## Version 0.1.0 (2022-03-19)
 
 Add vertical difference hashing.